+++ test/src/errors.js
@@ -11,14 +11,14 @@ describe('errors', function() {
         
     it("throws an error if the input bit depth is not valid", function () {
         testFunc = function() {
-            bitDepth([], "57", "16");
+            bitDepth([], "57", new Float64Array(), "16");
         };
         assert.throws(testFunc, /Invalid bit depth./);
     });
 
     it("throws an error if the output bit depth is not valid", function () {
         testFunc = function() {
-            bitDepth([], "16", "57");
+            bitDepth([], "16", new Float64Array(), "57");
         };
         assert.throws(testFunc, /Invalid bit depth./);
     });