+++ test/src/interface.js
@@ -11,43 +11,43 @@ describe("interface", function() {
     it("have the bitDepth function available", function() {
         const data = new Float64Array([1,-1]);
         var output = new Float64Array(data.length);
-        bitDepth(data, "8", "8", output);
+        bitDepth(data, "8", output, "8");
         assert.ok(output);
     });
     it("have the bitDepth function available", function() {
         const data = new Float64Array([1,-1]);
         var output = new Float64Array(data.length);
-        bitDepth(data, "16", "8", output);
+        bitDepth(data, "16", output, "8");
         assert.ok(output);
     });
     it("have the bitDepth function available", function() {
         const data = new Float64Array([1,-1]);
         var output = new Float64Array(data.length);
-        bitDepth(data, "32f", "8", output);
+        bitDepth(data, "32f", output, "8");
         assert.ok(output);
     });
     it("have the bitDepth function available", function() {
         const data = new Float64Array([1,-1]);
         var output = new Float64Array(data.length);
-        bitDepth(data, "8", "32f", output);
+        bitDepth(data, "8", output, "32f");
         assert.ok(output);
     });
     it("have the bitDepth function available", function() {
         const data = new Float64Array([1,-1]);
         var output = new Float64Array(data.length);
-        bitDepth(data, "32f", "64", output);
+        bitDepth(data, "32f", output, "64");
         assert.ok(output);
     });
     it("have the bitDepth function available", function() {
         const data = new Float64Array([1,-1]);
         var output = new Float64Array(data.length);
-        bitDepth(data, "32f", "64", output);
+        bitDepth(data, "32f", output, "64");
         assert.ok(output);
     });
-    it("throw an error if the output bit depth is not valid", function () {
-        testFunc = function() {
-            bitDepth([], "16", "57",[]);
-        };
-        assert.throws(testFunc, /Invalid bit depth./);
+    it("truncate fp samples when converting to int", function() {
+        const data = new Float64Array([300,-2]);
+        var output = new Float64Array(data.length);
+        bitDepth(data, "64", output, "8");
+        assert.deepEqual(output, new Float64Array([255, 0]));
     });
 });